==> 1_Constitution_database/1_Search_equation.ipynb <==
import pandas as pd

### Loading country database
Based on https://unstats.un.org/unsd/methodology/m49/overview/

UNSD_Path = "../0_Reference_files/UNSD_database.csv"
countries_df = pd.read_csv(UNSD_Path, encoding='utf-8')
countries_list = countries_df['Country0'].tolist()
demonym1_list = countries_df[countries_df['Demonym1'].notnull()].Demonym1.tolist()
demonym2_list = countries_df[countries_df['Demonym2'].notnull()].Demonym2.tolist()
geo_list = countries_list + demonym1_list + demonym2_list

### Equation on WOS

charact = 'TI=('
for country in geo_list[:-1]: 
    charact += '"'+ str(country)+ '" OR '
charact += '"' + str(geo_list[-1]) + '") AND '

charact += 'TS =(("mitigation" OR "low carbon" OR "decarboni*ation" OR (("carbon" OR "CO2" OR "GHG" OR "greenhouse gas" OR "emission*") NEAR/3 "reduc*")) AND (' 
for k in range (2025, 2100): 
    charact += '"'+ str(k)+ '" OR '
charact += ' "2100"))' 

charact

### Equation on Scopus 
As the metadata can be downloaded by batch of maximum 2000 publications, we cut the search query in three groups of countries/demonyms. 

charact_sco1 = 'TITLE('
for country in geo_list[:129]: 
    charact_sco1 += '"'+ str(country)+ '" OR '
charact_sco1 += '"' + str(geo_list[130]) + '") AND '

charact_sco1 += 'TITLE-ABS-KEY(("mitigation" OR "low carbon" OR "decarboni*ation" OR (("carbon" OR "CO2" OR "GHG" OR "greenhouse gas" OR "emission*") W/3 "reduc*")) AND (' 
for k in range (2025, 2100): 
    charact_sco1 += '"'+ str(k)+ '" OR '
charact_sco1 += ' "2100"))' 

len(geo_list)

charact_sco1

charact_sco2 = 'TITLE('
for country in geo_list[131:299]: 
    charact_sco2 += '"'+ str(country)+ '" OR '
charact_sco2 += '"' + str(geo_list[300]) + '") AND '

charact_sco2 += 'TITLE-ABS-KEY(("mitigation" OR "low carbon" OR "decarboni*ation" OR (("carbon" OR "CO2" OR "GHG" OR "greenhouse gas" OR "emission*") W/3 "reduc*")) AND (' 
for k in range (2025, 2100): 
    charact_sco2 += '"'+ str(k)+ '" OR '
charact_sco2 += ' "2100"))' 

charact_sco2

charact_sco3 = 'TITLE('
for country in geo_list[300:-2]: 
    charact_sco3 += '"'+ str(country)+ '" OR '
charact_sco3 += '"' + str(geo_list[-1]) + '") AND '

charact_sco3 += 'TITLE-ABS-KEY(("mitigation" OR "low carbon" OR "decarboni*ation" OR (("carbon" OR "CO2" OR "GHG" OR "greenhouse gas" OR "emission*") W/3 "reduc*")) AND (' 
for k in range (2025, 2100): 
    charact_sco3 += '"'+ str(k)+ '" OR '
charact_sco3 += ' "2100"))' 

charact_sco3
