==> 1_Constitution_database/2_Fusion_Scopus.ipynb <==
import pandas as pd 

### Function to concatenate the batchs of maximum 2000 metadata downloaded on Scopus
Argument *n* is the total number of downloaded batchs 

def fusion(n):
    mainpath= "./input/SCOPUS/scopus"
    frames =[]
    for k in range(n): 
        file = mainpath + str(k) +".csv"
        df = pd.read_csv(file)
        frames.append(df)  
    Df = pd.concat(frames, sort=False)    
    return(Df)

df = fusion(3)

print("%d X %d dataframe" % (len(df), len(df.columns) ))

### Eliminate duplicates based on Scopus ID (*EID*)

df_without_dupli = df.groupby(['EID']).last().reset_index()

print("%d X %d dataframe" % (len(df_without_dupli), len(df_without_dupli.columns) ))

### Rename and select columns of interest

df_without_dupli.rename(columns ={'EID': 'scopus_number',
                                  'Title': 'title',
                                  'Authors': 'authors',
                                  'Source title': 'source',
                                  'DOI': 'doi',
                                  'Year': 'publication_year',
                                  'Abstract': 'abstract',
                                  'Author Keywords': 'author_keywords',
                                  'Document Type': 'docu_type'}, inplace=True)

col=['scopus_number','title','authors','source','doi','publication_year','abstract','author_keywords','docu_type']
df_without_dupli = df_without_dupli.reindex(columns=col)
df_without_dupli.head()

print("%d X %d dataframe" % (len(df_without_dupli), len(df_without_dupli.columns) ))

df_without_dupli.to_csv("./intermed/df_Scopus.csv",index=False)
