==> 1_Constitution_database/2_Fusion_WOS.ipynb <==
import pandas as pd 
import numpy as np
pd.options.display.max_columns = 500

### Functions to concatenate the batchs of maximum 500 metadata downloaded on WoS <br> (respectively in the *All databases* and and *Core collection*)
Argument *n* is the total number of downloaded batchs in each folder

def fusion_tot(n):
    mainpath= "./input/WOS/All_databases/savedrecs"
    frames =[]
    for k in range(n): 
        file = mainpath + str(k) +".txt"
        df = pd.read_table(file, sep = '\t', header = 0, index_col=False)
        frames.append(df)  
    Df = pd.concat(frames, sort=False)    
    return(Df)

def fusion_core(n):
    mainpath= "./input/WOS/Core_collection/savedrecs"
    frames =[]
    for k in range(n): 
        file = mainpath + str(k) +".txt"
        df = pd.read_table(file, sep = '\t', header = 0, index_col=False)
        frames.append(df)  
    Df = pd.concat(frames, sort=False)    
    return(Df)

### Read the database listing all WoS publications

df = fusion_tot(8)
print("%d X %d dataframe" % (len(df), len(df.columns) ))

### Add a column for author keywords in the database listing all WoS publications

df["DE"] = np.nan

### Eliminate duplicates based on WoS ID (*UT*)

df_key_null = df.groupby(['UT']).first().reset_index()
print("%d X %d dataframe" % (len(df_key_null), len(df_key_null.columns) ))

### Read the specific databases 

df1 = fusion_core(8)
print("%d X %d dataframe" % (len(df1), len(df1.columns) ))

df2 = pd.read_table("./input/WOS/Current_contents_connect/savedrecs0.txt", sep = '\t', header = 0, index_col=False)
df3 = pd.read_table("./input/WOS/KCI-Korean/savedrecs0.txt", sep = '\t', header = 0, index_col=False)
df4 = pd.read_table("./input/WOS/MEDLINE/savedrecs0.txt", sep = '\t', header = 0, index_col=False)
df5 = pd.read_table("./input/WOS/Russian_Science/savedrecs0.txt", sep = '\t', header = 0, index_col=False)
df6 = pd.read_table("./input/WOS/SciELO/savedrecs0.txt", sep = '\t', header = 0, index_col=False)

### Index all databases with the WoS ID and update the database listing all publications with information in specific databases

df.set_index("UT", inplace = True)
df1.set_index("UT", inplace = True)
df2.set_index("UT", inplace = True)
df3.set_index("UT", inplace = True)
df4.set_index("UT", inplace = True)
df5.set_index("UT", inplace = True)
df6.set_index("UT", inplace = True)

df.update(df1)
df.update(df2)
df.update(df3)
df.update(df4)
df.update(df5)
df.update(df6)

df.reset_index(0, inplace =True)

df.head(1)

### Rename and select columns of interest

df.rename(columns ={'UT': 'WOS_number',
                    'TI': 'title',
                    'PT': 'doc_type',
                    'AU': 'authors',
                    'SO': 'source',
                    'DI': 'doi',
                    'PY': 'publication_year',
                    'AB': 'abstract',
                    'DE': 'author_keywords'}, inplace=True)

col=['WOS_number','doc_type','authors','title','source','doi','publication_year','abstract','author_keywords','email']
df = df.reindex(columns=col)
df.head()

print("%d X %d dataframe" % (len(df), len(df.columns) ))

df.to_csv("./intermed/df_WOS.csv",index=False)
